*Ensured calls to snprintf are passed 1 byte less than the buffer's length, and buffers used were initialized to NULL, preventing the occurence of non-terminated strings. *Casted unused return values to (void), which is a nit-picky splint recommendation, but it does allow the intended handling of the value to be explicitly shown. *Changed various variable data types and casted throughout to satisfy splint.  At this point, it appears most of the meaningful warnings have been resolved. *Fixed some wild memory freeing in some of the "Erase all" functions, which could plausibly have caused a crash.

+++ trunk/src/event.c
@@ -124,7 +124,7 @@ void eof_sort_events(EOF_SONG * sp)
 
 	if(sp)
 	{
-		qsort(sp->text_event, sp->text_events, sizeof(EOF_TEXT_EVENT *), eof_song_qsort_events);
+		qsort(sp->text_event, (size_t)sp->text_events, sizeof(EOF_TEXT_EVENT *), eof_song_qsort_events);
 	}
 }
